+++ app/src/components/loader.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import styled from 'styled-components';
+
 import CircularProgress from '@material-ui/core/CircularProgress';
+
+export const SmallLoader = () => (
+    <CircularProgress />
+);
 
 const Loader = () => (
     <LoaderContainer>
@@ -8,14 +13,9 @@ const Loader = () => (
     </LoaderContainer>
 );
 
-export const SmallLoader = () => (
-    <CircularProgress />
-);
-
 export default Loader;
 
 const LoaderContainer = styled.div`
     display: flex;
     justify-content: center;
-    /* padding-top: 25vh; */
 `;
