+++ app/src/container/footer/footer.js
@@ -1,10 +1,10 @@
-import React, { useState } from 'react';
-import BottomNavigation from '@material-ui/core/BottomNavigation';
+import React from 'react';
+
 import BottomNavigationAction from '@material-ui/core/BottomNavigationAction';
 import SearchTwoToneIcon from '@material-ui/icons/SearchTwoTone';
 import ContactMailTwoToneIcon from '@material-ui/icons/ContactMailTwoTone';
 
-import styled from 'styled-components';
+import { StyledSearch, StyledBottomNavigation } from './footer.css.js';
 
 const Footer = ({ onPageChange }) => {
     const [value, setValue] = React.useState('search');
@@ -23,20 +23,3 @@ const Footer = ({ onPageChange }) => {
 };
 
 export default Footer;
-
-const StyledSearch = styled.div`
-    position: fixed;
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    text-align: center;
-`;
-
-const StyledBottomNavigation = styled(BottomNavigation)`
-    && {
-        background-color: #424242;
-    }
-    && button > span {
-        color: white;
-    }
-`;
