+++ app/src/container/header/header.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 
-import AppBar from '@material-ui/core/AppBar';
+import { HeaderText, StyledAppBar, AppBarContainer } from './header.css.js';
 
 import COMMON_TEXT from '../../assets/wording/common.json';
 
@@ -16,20 +15,3 @@ const Header = ({ minwidth }) => {
 };
 
 export default Header;
-
-const HeaderText = styled.div`
-    font-size: 18px;
-    padding: 15px; 
-`;
-
-const StyledAppBar = styled(AppBar)`
-    min-width: ${({minwidth}) => `${minwidth}px`};
-`;
-
-const AppBarContainer = styled.div`
-    && {
-        background-color: #424242;
-    }
-    display: flex;
-    justify-content: space-between;
-`;
