+++ app/src/container/main.js
@@ -6,6 +6,7 @@ import Header from './header/header';
 import Footer from './footer/footer';
 import Search from './search/search';
 import Contact from './contact/contact';
+import { MainContainer } from './style.css.js';
 
 import { sendRequest, Api } from '../utils/httpService';
 import { STYLE } from '../config/common';
@@ -39,7 +40,7 @@ const Main = () => {
     return (
         <>
             <Header minwidth={STYLE.MIN_WIDTH} />
-            <Container>
+            <MainContainer>
                 {
                     page === PAGES.SEARCH && (
                         <Search
@@ -51,15 +52,10 @@ const Main = () => {
                 {
                     page === PAGES.CONTACT && (<Contact />)
                 }
-            </Container>
+            </MainContainer>
             <Footer onPageChange={setPage} />
         </>
     );
 };
 
 export default Main;
-
-const Container = styled.div`
-    min-width: ${STYLE.MIN_WIDTH}px;
-    padding: 64px 0px;
-`;
