+++ app/src/container/search/search.js
@@ -1,30 +1,39 @@
 // core
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 
 // third party component
 import Fuse from 'fuse.js';
 import Button from '@material-ui/core/Button';
-import SearchIcon from '@material-ui/icons/Search';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
-import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import ChipInput from 'material-ui-chip-input';
 
-// component
+// components
 import Loader from '../../components/loader';
 import {
-    Container,
-    Header,
-    Item,
-    Content
-} from '../style.css.js';
+    SwitchText,
+    SearchContainer,
+    SearchIndicator,
+    StyledSearchIcon,
+    Cards,
+    CardTitle,
+    ItemName,
+    IGPostCard,
+    Image,
+    StyledExpansionPanel,
+    OuterExpansionPanel,
+    OuterExpansionPanelSummary,
+    OuterExpansionPanelDetails,
+    InnerExpansionPanelSummary,
+    InnerExpansionPanelDetails,
+    StyledRecommendBlock,
+    ButtonGroup
+} from './search.css.js';
+import { Container, Header, Item, Content } from '../style.css.js';
 
-// config and assets
-import { STYLE } from '../../config/common';
+// utils, config and assets
+import { findExistInTwoArray } from '../../utils/helper';
 import { RECOMMEND } from '../../config/search';
 import SEARCH_TEXT from '../../assets/wording/search.json';
 
@@ -37,15 +46,6 @@ const DEFAULT_STATE = {
     searchResultList: [],
     showRecommend: true,
     expanded: '' 
-};
-
-// helper
-const findExistInTwoArray = (array1, array2) => {
-    for (let i = 0; i < array1.length; i += 1) {
-        const find = array2.findIndex(elm => elm.toLowerCase() === array1[i].toLowerCase()) > -1
-        if (find) return find;
-    }
-    return false;
 };
 
 const RecommendBlock = ({ dataObj, expanded, onItemSelect, onExpanded, showRecommend, onControlRecommend }) => {
@@ -69,9 +69,9 @@ const RecommendBlock = ({ dataObj, expanded, onItemSelect, onExpanded, showRecom
                                     onChange={() => { onExpanded(title)} }
                                     key={`item-${title}`}
                                 >
-                                    <ExpansionPanelSummary expandIcon={<ExpandMoreIcon />}>
+                                    <InnerExpansionPanelSummary expandIcon={<ExpandMoreIcon />}>
                                         <div>{title}</div>
-                                    </ExpansionPanelSummary>
+                                    </InnerExpansionPanelSummary>
                                     <InnerExpansionPanelDetails>
                                         <ButtonGroup>
                                             {
@@ -287,112 +287,3 @@ const Search = ({ searchPageData, getCocktailsList }) => {
 };
 
 export default Search;
-
-const SwitchText = styled.span`
-    font-size: 14px;
-    color: ${({checked}) => checked ? 'black' : 'grey'};
-`;
-
-const SearchContainer = styled.div`
-    display: flex;
-    width: 300px;
-    margin-right: ${STYLE.PADDING}px;
-`;
-
-const SearchIndicator = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-`;
-
-const StyledSearchIcon = styled(SearchIcon)`
-    height: 48px;
-`;
-
-const Cards = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-`;
-
-const CardTitle = styled.div`
-    display: flex;
-    width: 270px;
-    text-align: center;
-    font-size: 1rem;
-    padding: 15px;
-    && > button {
-        margin-left: 10px;
-    }
-`;
-
-const ItemName = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    width: 190px;
-`;
-
-const IGPostCard = styled.div`
-    margin: 10px;
-    padding-bottom: 15px;
-    background-color: white;
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-    transition: 0.3s;
-    &&:hover {
-        box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-    }
-`;
-
-const Image = styled.div`
-    width: 300px;
-    height: 300px;
-    background-size: 100%;
-    background-position: center;
-    background-repeat: no-repeat;
-    background-image: url(${({ url }) => url});
-    &&:hover {
-        cursor: pointer;
-    }
-`;
-
-const StyledExpansionPanel = styled(ExpansionPanel)`
-    width: 320px;
-    && button {
-        margin: 5px;
-    }
-`;
-
-const OuterExpansionPanel = styled(ExpansionPanel)`
-    width: ${STYLE.MIN_WIDTH}px;
-    &&& {
-        background-color: #f4f4ec;
-        box-shadow: none;
-        margin-right: 20px;
-    }
-`;
-
-const OuterExpansionPanelSummary = styled(ExpansionPanelSummary)`
-    width: ${STYLE.MIN_WIDTH}px;
-`;
-
-const OuterExpansionPanelDetails = styled(ExpansionPanelDetails)`
-    display: flex;
-    flex-direction: column;
-`;
-
-const InnerExpansionPanelDetails = styled(ExpansionPanelDetails)`
-    && {
-        padding: ${STYLE.PADDING}px 15px;
-    }
-`;
-
-const StyledRecommendBlock = styled.div`
-    display: flex;
-    flex-direction: column;
-`;
-
-const ButtonGroup = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-`;
